requirements.txt
anyjson==0.3.3
appdirs==1.4.4
APScheduler==3.6.3
arrow==0.15.6
asgiref==3.2.10
billiard==3.6.3.0
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cookiecutter==1.7.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.2
django-filter==2.4.0
django-timezone-field==4.0
djangorestframework==3.12.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
Jinja2==2.11.2
jinja2-time==0.2.0
kombu==5.0.2
Markdown==3.3.1
MarkupSafe==1.1.1
pipenv==2020.6.2
poyo==0.5.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tzlocal==2.1
urllib3==1.25.9
vine==5.0.0
virtualenv==20.0.23
virtualenv-clone==0.5.4
wcwidth==0.2.5
zipp==3.1.0
